add mergeKeyValues patch hook
patch will normally replace the keyValues map w/ the new one, but we want to merge
the keys in the patch with the existing keys.

Perhaps we want something more sophisticated such as an op param to specify merge/replace or remove?
Also we probably should verify that the data.keyValues Map exists.